Add implementation for ClassResourceController

+++ src/main/java/org/caskol/warcraft_database/api/v1/services/implementations/ClassResourceServiceImpl.java
@@ -1,37 +1,74 @@
 package org.caskol.warcraft_database.api.v1.services.implementations;
 
+import jakarta.validation.ValidationException;
+import lombok.Getter;
 import lombok.RequiredArgsConstructor;
+import lombok.Setter;
 import org.caskol.warcraft_database.api.v1.dto.ClassResourceDTO;
+import org.caskol.warcraft_database.api.v1.exceptions.NoSuchElementFoundException;
+import org.caskol.warcraft_database.api.v1.mappers.ClassResourceMapper;
+import org.caskol.warcraft_database.api.v1.models.ClassResource;
 import org.caskol.warcraft_database.api.v1.repositories.ClassResourceRepository;
 import org.caskol.warcraft_database.api.v1.services.ClassResourceService;
+import org.caskol.warcraft_database.utils.RestExceptionHandler;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.validation.Errors;
+import org.springframework.validation.Validator;
 
 import java.util.List;
 import java.util.Optional;
 
-
+@Getter
+@Setter
 @Service
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class ClassResourceServiceImpl implements ClassResourceService {
     private final ClassResourceRepository classResourceRepository;
-    public Optional<ClassResourceDTO> getById(int id) {
-        return null;
-        //return classResourceRepository.findById(id);
+    private final ClassResourceMapper classResourceMapper;
+    private final Validator validator;
+
+    @Transactional(readOnly = false)
+    public ClassResourceDTO create(ClassResourceDTO classResourceDTO) {
+        ClassResource newClassResource = classResourceMapper.toClassResource(classResourceDTO);
+        classResourceRepository.save(newClassResource);
+        return classResourceMapper.allDataToDTO(newClassResource);
+    }
+
+    public ClassResourceDTO getById(int id)
+    {
+        Optional<ClassResource> classResource = classResourceRepository.findById(id);
+        if (classResource.isPresent()) {
+            return classResourceMapper.allDataToDTO(classResource.get());
+        }
+        else
+            throw new NoSuchElementFoundException(ClassResource.class.getSimpleName()+ " with id="+id+" was not found");
     }
     @Transactional(readOnly = false)
-    public void save(ClassResourceDTO classResourceDTO) {
-        //classResourceRepository.save(classResource);
+    public void update(ClassResourceDTO classResourceDTO)
+    {
+        Optional<ClassResource> iconFromRepo = classResourceRepository.findById(classResourceDTO.getId());
+        ClassResource classResource = iconFromRepo.orElseThrow(() -> new NoSuchElementFoundException(ClassResource.class.getSimpleName()+ " with id="+classResourceDTO.getId()+" was not found"));
+        classResourceMapper.updateClassResourceFromDto(classResourceDTO,classResource);
+        Errors errors = validator.validateObject(classResource);
+        if (errors.hasErrors())
+            throw new ValidationException(RestExceptionHandler.VALIDATION_EXCEPTION_MSG + RestExceptionHandler.getValidationErrorString(errors));
+        classResourceRepository.save(classResource);
     }
     @Transactional(readOnly = false)
     public void delete(int id)
     {
+        if (!classResourceRepository.findById(id).isPresent())
+            throw new NoSuchElementFoundException(ClassResource.class.getSimpleName()+ " with id="+ id+ " was not found");
         classResourceRepository.deleteById(id);
     }
     public List<ClassResourceDTO> getAll()
     {
-        return null;
-        //return classResourceRepository.findAll();
+        return classResourceRepository.findAll()
+                .stream()
+                .map(classResourceMapper::dataWithoutListToDTO)
+                .toList();
     }
+
 }
